@@ -1,22 +1,23 @@
-/**
- * This file is part of morfi which is released under MIT license.
- *
- * The LICENSE file can be found in the root directory of this project.
- *
- * @flow
- */
+// @flow
 
-import React, { Component } from 'react';
+import React, { useCallback, useRef, useState } from 'react';
 import { Validators } from '../../validators/validators';
-import { Form } from '../../form/index';
-import { FormInput } from '../../fields/index';
-import type { FormData, FormValidation, Validator } from '../../form/index';
 import { Spinner } from '../../icons/Spinner';
+import { FormInput } from '../../fields/FormInput';
+import { Morfi, type FormData, type FormValidation } from '../../form';
+import { useSafeState } from '../../../hooks/useSafeState';
 
-type FormValues = {
-    userName?: string,
-    realName?: string,
-};
+// some fake server state
+const alreadyRegistered_userName = 'tom';
+const everySecondTimeFailing_userName = 'jack';
+const failingAfterSubmit_userName = 'mike';
+let jackValidationCounter = 0;
+
+type FormValues = {| userName: string, realName: string, alias: string |};
+
+const initialValues: FormValues = { userName: '', realName: '', alias: '' };
+
+const { Form, Fields } = Morfi.create(initialValues);
 
 const StaticValidations: FormValidation<FormValues> = {
     realName: {
@@ -29,122 +30,108 @@ const StaticValidations: FormValidation<FormValues> = {
     },
 };
 
-type AsyncValidationSampleState = {
-    data: FormData<FormValues>,
-    validations: FormValidation<FormValues>,
-    pendingUserName: boolean,
+const onSubmit = ({ userName = '' }: FormValues): Promise<void> => {
+    // simulate server request
+    const fakeServerRequest = window.sleep(1000);
+    return userName.toLowerCase() !== failingAfterSubmit_userName
+        ? fakeServerRequest
+        : fakeServerRequest.then(() => {
+              throw new Error('AsyncValidationSample.userName.submit.failed');
+          });
 };
 
-const alreadyRegistered_userName = 'tom';
-const everySecondTimeFailing_userName = 'jack';
-const failingAfterSubmit_userName = 'mike';
-let validationCounter = 0;
-
-export default class AsyncValidationSample extends Component<{}, AsyncValidationSampleState> {
-    validateUserName: Validator = (userName?: string) => {
-        const syncError = Validators.string({ min: 1 })(userName);
-        if (syncError) {
-            return syncError;
+const validateName = (setPending: boolean => void) => (name?: string) => {
+    const syncError = Validators.string({ min: 1 })(name);
+    if (syncError) {
+        return syncError;
+    }
+    const lowerCaseUserName = (name: any).toLowerCase();
+    setPending(true);
+    lowerCaseUserName === everySecondTimeFailing_userName && jackValidationCounter++;
+    return window.sleep(1000).then(() => {
+        setPending(false);
+        if (lowerCaseUserName === alreadyRegistered_userName) {
+            return { id: 'AsyncValidationSample.userName.already.registered', values: { userName: name } };
+        } else if (lowerCaseUserName === everySecondTimeFailing_userName && jackValidationCounter % 2 === 0) {
+            return { id: 'AsyncValidationSample.userName.already.registered', values: { userName: name } };
         }
-        const lowerCaseUserName = (userName: any).toLowerCase();
-        this.setState({ pendingUserName: true });
-        validationCounter++;
-        return window.sleep(1000).then(() => {
-            if (lowerCaseUserName === alreadyRegistered_userName) {
-                this.setState({ pendingUserName: false });
-                return { id: 'AsyncValidationSample.userName.already.registered', values: { userName } };
-            } else if (lowerCaseUserName === everySecondTimeFailing_userName && validationCounter % 2 === 0) {
-                this.setState({ pendingUserName: false });
-                return { id: 'AsyncValidationSample.userName.already.registered', values: { userName } };
-            } else {
-                this.setState({ pendingUserName: false });
-            }
-        });
-    };
+    });
+};
 
-    getValidations = (): FormValidation<FormValues> => {
-        return {
-            ...StaticValidations,
-            userName: { onBlur: this.validateUserName, onChange: Validators.string({ min: 1 }) },
-        };
-    };
+export const AsyncValidationSample = () => {
+    const [pendingUserName, setPendingUserName] = useSafeState<boolean>(false);
+    const validateUserName = useCallback(validateName(setPendingUserName), []);
 
-    state: AsyncValidationSampleState = {
-        data: { values: {}, errors: {} },
-        pendingUserName: false,
-        validations: this.getValidations(),
-    };
+    const [pendingAlias, setPendingAlias] = useSafeState<boolean>(false);
+    const validateAlias = useCallback(validateName(setPendingAlias), []);
 
-    onChange = (data: FormData<FormValues>) => this.setState({ data });
+    const validation = useRef<FormValidation<FormValues>>({
+        ...StaticValidations,
+        userName: { onBlur: validateUserName, onChange: Validators.string({ min: 1 }) },
+        alias: { onChange: validateAlias },
+    });
 
-    onSubmit = ({ userName = '' }: FormValues): Promise<void> => {
-        // simulate server request
-        const fakeServerRequest = window.sleep(1000);
-        return userName.toLowerCase() !== failingAfterSubmit_userName
-            ? fakeServerRequest
-            : fakeServerRequest.then(() => {
-                  throw new Error('AsyncValidationSample.userName.submit.failed');
-              });
-    };
+    const [data, setData] = useState<FormData<FormValues>>({ values: initialValues, errors: {} });
 
-    onSubmitFailed = (e: Error): void => {
-        Form.isValidationError(e) ||
-            this.setState({
-                data: {
-                    values: { ...this.state.data.values },
-                    errors: { ...this.state.data.errors, userName: { id: e.message } },
-                },
-            });
-    };
+    const onSubmitFailed = useCallback((e: Error): void => {
+        Morfi.isValidationError(e) ||
+            setData(({ values, errors }) => ({
+                values: { ...values },
+                errors: { ...errors, userName: { id: e.message } },
+            }));
+    }, []);
 
-    onSubmitFinished = (): void => {
-        this.setState({ data: { values: {}, errors: {} } });
-    };
+    const onSubmitFinished = useCallback((): void => {
+        setData({ values: initialValues, errors: {} });
+    }, []);
 
-    render(): React$Node {
-        const { data, pendingUserName } = this.state;
-        const { values, errors, submitting } = data;
-        return (
-            <div className="col-12">
-                <Form
-                    validation={this.state.validations}
-                    onChange={this.onChange}
-                    data={data}
-                    onSubmit={this.onSubmit}
-                    onSubmitFailed={this.onSubmitFailed}
-                    onSubmitFinished={this.onSubmitFinished}>
-                    <div className="row">
-                        <div className="col-md-12">
-                            <FormInput
-                                name="userName"
-                                value={values.userName}
-                                error={errors.userName}
-                                pending={pendingUserName}
-                                label="Username"
-                                placeholder="Please enter your desired username"
-                            />
-                            <p className="small font-italic form-group-apply">
-                                ATTENTION: The validation succeeds for all names, but the following:<br />
-                                <strong>Tom</strong>: Validation fails each time.<br />
-                                <strong>Jack</strong>: Validation fails each second time.<br />
-                                <strong>Mike</strong>: Validation succeeds but submitting fails.
-                            </p>
-                            <FormInput
-                                name="realName"
-                                value={values.realName}
-                                error={errors.realName}
-                                label="Real name"
-                                placeholder="Please enter your real name"
-                            />
-                            <div className="btn-toolbar">
-                                <button className="btn btn-success" disabled={submitting || Form.hasErrors(data)}>
-                                    {submitting && <Spinner />} Submit
-                                </button>
-                            </div>
+    const { submitting } = data;
+
+    return (
+        <div className="col-12">
+            <Form
+                validation={validation.current}
+                onChange={setData}
+                data={data}
+                onSubmit={onSubmit}
+                onSubmitFailed={onSubmitFailed}
+                onSubmitFinished={onSubmitFinished}>
+                <div className="row">
+                    <div className="col-md-12">
+                        <p className="small font-italic form-group-apply">
+                            ATTENTION: The validation succeeds for all names, but the following:
+                            <br />
+                            <strong>Tom</strong>: Validation fails each time.
+                            <br />
+                            <strong>Jack</strong>: Validation fails each second time.
+                            <br />
+                            <strong>Mike</strong>: Validation succeeds but submitting fails.
+                        </p>
+                        <FormInput
+                            Field={Fields.userName}
+                            pending={pendingUserName}
+                            label="Username (async validation triggers on blur)"
+                            placeholder="Please enter your desired username"
+                        />
+                        <FormInput
+                            Field={Fields.alias}
+                            pending={pendingAlias}
+                            label="Alias (async validation triggers on change)"
+                            placeholder="Please enter some alias"
+                        />
+                        <FormInput
+                            Field={Fields.realName}
+                            label="Real name"
+                            placeholder="Please enter your real name"
+                        />
+                        <div className="btn-toolbar">
+                            <button className="btn btn-success" disabled={submitting || Morfi.hasErrors(data)}>
+                                {submitting && <Spinner />} Submit
+                            </button>
                         </div>
                     </div>
-                </Form>
-            </div>
-        );
-    }
-}
+                </div>
+            </Form>
+        </div>
+    );
+};
