@@ -12,9 +12,6 @@ describe('Validator.string', () => {
     it('does not process falsy values to the validator on optionalOf-variant', () => {
         const validator = Validators.optionalOf(Validators.string({ min: 3, max: 10 }));
         expect(validator(undefined)).toBe(undefined);
-        expect(validator(null)).toBe(undefined);
-        expect(validator(12)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(true)).toEqual({ id: 'validation.value.incompatible' });
         expect(validator('')).toBe(undefined);
         expect(validator('some non empty string')).toEqual({
             id: 'validation.characters.atMost.x',
@@ -25,9 +22,6 @@ describe('Validator.string', () => {
     it('validates strings correctly', () => {
         const validator = Validators.string({});
         expect(validator(undefined)).toBe(undefined);
-        expect(validator(null)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(12)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(true)).toEqual({ id: 'validation.value.incompatible' });
         expect(validator('')).toBe(undefined);
         expect(validator('some non empty string')).toBe(undefined);
     });
@@ -60,18 +54,12 @@ describe('Validator.number', () => {
     it('validates numbers correctly', () => {
         const validator = Validators.number({});
         expect(validator(undefined)).toBe(undefined);
-        expect(validator(null)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(true)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator('123A')).toEqual({ id: 'validation.value.onlyNumbers' });
         expect(validator(12)).toBe(undefined);
-        expect(validator('12')).toBe(undefined);
-        expect(validator('')).toBe(undefined); // because of Number('') === 0, we want to explicitly allow empty string
     });
 
     it('validates number min length correctly', () => {
         const validator = Validators.number({ minLength: 3 });
         const expectedError = { id: 'validation.characters.atLeast.x', values: { num: 3, s: 's' } };
-        expect(validator('')).toEqual(expectedError);
         expect(validator(1)).toEqual(expectedError);
         expect(validator(12)).toEqual(expectedError);
         expect(validator(123)).toBe(undefined);
@@ -82,7 +70,6 @@ describe('Validator.number', () => {
         const validator = Validators.number({ maxLength: 4 });
         const expectedError = { id: 'validation.characters.atMost.x', values: { num: 4, s: 's' } };
         expect(validator(undefined)).toBe(undefined);
-        expect(validator('')).toBe(undefined);
         expect(validator(1)).toBe(undefined);
         expect(validator(12)).toBe(undefined);
         expect(validator(123)).toBe(undefined);
@@ -93,14 +80,14 @@ describe('Validator.number', () => {
 
     it('special behaviour for min length = 1', () => {
         const validatorMin1 = Validators.number({ minLength: 1 });
-        expect(validatorMin1('')).toEqual({ id: 'validation.characters.atLeast.x', values: { num: 1, s: '' } });
+        expect(validatorMin1(undefined)).toEqual({ id: 'validation.characters.atLeast.x', values: { num: 1, s: '' } });
         expect(validatorMin1(0)).toEqual(undefined);
     });
 
     it('special behaviour for max length = 0', () => {
         const validatorMax1 = Validators.number({ maxLength: 0 });
         expect(validatorMax1(0)).toEqual({ id: 'validation.characters.exactly.x', values: { num: 0, s: 's' } });
-        expect(validatorMax1('')).toEqual(undefined);
+        expect(validatorMax1(undefined)).toEqual(undefined);
     });
 });
 
@@ -110,9 +97,6 @@ describe('Validator.regex', () => {
     it('validates regex correctly', () => {
         const validator = Validators.regex({ re: regex });
         expect(validator(undefined)).toEqual({ id: 'validation.value.mismatch' });
-        expect(validator(null)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(true)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(12)).toEqual({ id: 'validation.value.incompatible' });
         expect(validator('does not match')).toEqual({ id: 'validation.value.mismatch' });
         expect(validator('(010101)')).toBe(undefined);
         expect(validator('(0101012)')).toEqual({ id: 'validation.value.mismatch' });
@@ -129,9 +113,6 @@ describe('Validator.email', () => {
     it('validates email correctly', () => {
         const validator = Validators.email;
         expect(validator(undefined)).toEqual({ id: 'validation.email.requirements' });
-        expect(validator(null)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(true)).toEqual({ id: 'validation.value.incompatible' });
-        expect(validator(12)).toEqual({ id: 'validation.value.incompatible' });
         expect(validator('some?non@sense')).toEqual({ id: 'validation.email.requirements' });
         expect(validator('1.matching-mail@address.com')).toBe(undefined);
         expect(validator('non-matching_address.com')).toEqual({ id: 'validation.email.requirements' });
